Created new Firmware
Changes most of how things work, can't comment on sensor compatibility yet but should be easier to understand. Encoder doesn't register yet and for some reason pressing the button takes two readings instead of one but it's a proof of concept that we can build on.
Still don't know exactly how to do the averaging but shouldn't be hard, and this uses full refreshes for everything but it's fast enough. This might need to change for the encoder though.